refactor: move floating-utils.ts to shared/
